pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "openenv-fraud"
version = "1.0.0"
description = "Fraud Detection & Transaction Risk Review OpenEnv Environment"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "OpenEnv Fraud Team" },
]
keywords = ["reinforcement-learning", "fraud-detection", "openenv", "fintech", "rl"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "License :: OSI Approved :: MIT License",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
  "openenv-core>=0.1.0",
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.29.0",
  "pydantic>=2.0.0",
  "numpy>=1.26.0",
  "httpx>=0.27.0",
  "python-dotenv>=1.0.0",
  "rich>=13.7.0",
  "openai>=1.30.0",
  "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.4.0",
  "mypy>=1.10.0",
]

[project.scripts]
server = "server.app:main"
fraud-server = "server.app:app"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.mypy]
python_version = "3.10"
strict = false
ignore_missing_imports = true
